requirements.txt
discord-py
termcolor
